+++ src/main/java/jak/dut/Investment/model/calculation/EndInvestmentCalculation.java
@@ -7,13 +7,13 @@ import java.time.LocalDate;
 public class EndInvestmentCalculation implements CalculaltionAlgorithmStrategy {
 
     @Override
-    public BigDecimal calculation(Calculation calculation) {
-        BigDecimal value = calculation.getAmount();
+    public Double calculation(Calculation calculation) {
+        Double value = calculation.getAmount();
         Double interest = calculation.getInvestment().getRate();
         LocalDate startDate = calculation.getInvestment().getStartInvestmentDate();
         LocalDate endDate = calculation.getInvestment().getEndInvestmentDate();
         Duration periodInDays = Duration.between(endDate,startDate);
-        BigDecimal finalIntrest;
+        Double finalIntrest;
         //double f=p*Math.pow((1+r/100),y);
 //        finalIntrest = value*Math.pow((1+interest),periodInDays)
         return null;
